splitting: make temporal private
